Add support for `aiida-pseudo`
This new library provides functionality to make it easier to install
pseudo potential families, either from an archive of pseudo potential
files on the local file system, or by directly downloading from the web.
There is special functionality to install an SSSP family which will also
install the corresponding recommended cutoffs that can be easily
retrieved from a family for a given set of elements or a structure.

For now, the library is just plumbed into the CLI commands that launch a
process that requires pseudos in the input. The plugin also provides a
new implementation of `UpfData` that is different from the one that
ships with `aiida-core` so the base calculation is updated to accept
both classes for the `pseudos` input namespace. They should work the
same from the process' perspective. At some point the data plugin from
the `aiida-pseudo` plugin should supersede that of `aiida-core` which
will be deprecated.